+++ config/connection.js
@@ -1,0 +1,31 @@
+require("dotenv").config();
+
+const Sequelize = require("sequelize");
+
+const sequelize = process.env.JAWSDB_URL
+  ? new Sequelize(process.env.JAWSDB_URL)
+  : new Sequelize(
+      process.env.DB_NAME,
+      process.env.DB_USER,
+      process.env.DB_PASSWORD,
+      {
+        host: "localhost",
+        dialect: "mysql",
+        dialectOptions: {
+          decimalNumbers: true,
+        },
+      }
+  );
+    
+
+const checkAuthenticity = async () => {
+  try {
+    await sequelize.authenticate();
+    console.log("Connection has been established successfully.");
+  } catch (error) {
+    console.error("Unable to connect to the database:", error);
+  }
+};
+
+checkAuthenticity();  
+module.exports = sequelize;
